IC-25 fix: add layout
