+++ src/app/(public)/(auth)/layout.tsx
@@ -1,0 +1,9 @@
+import s from './styles.module.scss'
+
+export default function Layout({
+  children,
+}: Readonly<{
+  children: React.ReactNode
+}>) {
+  return <div className={s.container}>{children}</div>
+}
