+++ src/app/(public)/(auth)/signin/page.tsx
@@ -1,4 +1,4 @@
-import { SignInPage } from '@/pageComponents/signIn'
+import { SignInPage } from '@/pagesComponents/signIn'
 
 export default function SignIn() {
   return <SignInPage />
